refactor(): * fixed all remaining dependency issues between packages * added a postinstall script to make dealing with examples in the web UI easier (just copied the data from the examples repo into an examples folder) * cleanups * added lerna config
