+++ lerna.json
@@ -1,0 +1,12 @@
+{
+  "lerna": "2.7.0",
+  "packages": [
+    "packages/*"
+  ],
+  "version": "independent",
+  "command": {
+    "init": {
+      "exact": true
+    }
+  }
+}
