+++ package.json
@@ -38,35 +38,11 @@
   ],
   "license": "MIT",
   "dependencies": {
-    "@jscad/csg": "0.3.6",
-    "@jscad/io": "0.3.7",
-    "@jscad/openscad-openjscad-translator": "0.0.10",
-    "@jscad/scad-api": "^0.5.1",
-    "astring": "^1.0.2",
-    "brace": "0.10.0",
-    "esprima": "^3.1.3",
-    "estraverse": "^4.2.0",
-    "most-gestures": "^0.3.0",
-    "webworkify": "^1.4.0"
   },
   "devDependencies": {
     "ava": "^0.23.0",
-    "babel-cli": "^6.6.5",
-    "babel-core": "^6.2.1",
-    "babel-preset-es2015": "^6.1.18",
-    "babelify": "^7.2.0",
-    "browserify": "14.3.0",
-    "browserify-shim": "^3.8.12",
-    "budo": "10.0.3",
-    "lerna": "^2.7.0",
-    "minifyify": "^7.3.3"
+    "lerna": "2.7.0"
   },
-  "browserify": {
-    "transform": [
-      "browserify-shim"
-    ]
-  },
-  "browserify-shim": {},
   "collective": {
     "type": "opencollective",
     "url": "https://opencollective.com/openjscad",
