+++ packages/cli/package.json
@@ -3,12 +3,11 @@
   "version": "1.6.1",
   "description": "",
   "repository": "https://github.com/jscad/OpenJSCAD.org",
-  "main": "src/module.js",
   "bin": {
-    "openjscad": "src/cli/cli.js"
+    "openjscad": "./cli.js"
   },
   "scripts": {
-    "test": "ava './src/cli/*.test.js' --verbose --timeout 10000",
+    "test": "ava './*.test.js' --verbose --timeout 10000",
     "build-all": "npm run build-web && npm run build-min && npm run build-opt",
     "preversion": "npm test",
     "version": "git add -A ",
@@ -34,10 +33,12 @@
   ],
   "license": "MIT",
   "dependencies": {
+    "@jscad/core": "1.6.1",
     "@jscad/csg": "0.3.6",
     "@jscad/io": "0.3.7",
     "@jscad/openscad-openjscad-translator": "0.0.10",
     "@jscad/scad-api": "^0.5.1",
+    "@jscad/examples": "1.6.1",
     "astring": "^1.0.2",
     "esprima": "^3.1.3",
     "estraverse": "^4.2.0"
