+++ packages/core/core/convertToSolid.js
@@ -1,6 +1,6 @@
 const { CSG, CAG } = require('@jscad/csg')
 const {toArray} = require('../utils/misc')
-const {isCSG, isCAG} = require('./utils')
+const {isCSG, isCAG} = require('../utils/isCSGCAG')
 
 // FIXME: is there not too much overlap with convertToBlob ?
 /**
