+++ packages/core/utils/isCSGCAG.js
