+++ packages/web/package.json
@@ -17,7 +17,7 @@
     "release-patch": "git checkout master && git pull origin master && npm version patch",
     "release-minor": "git checkout master && git pull origin master && npm version minor",
     "release-major": "git checkout master && git pull origin master && npm version major",
-    "postinstall": "node -e \"console.log('\\u001b[35m\\u001b[1mLove OpenJSCAD? You can now donate to our open collective:\\u001b[22m\\u001b[39m\\n > \\u001b[34mhttps://opencollective.com/openjscad/donate\\u001b[0m')\""
+    "postinstall": "node postinstall.js && node -e \"console.log('\\u001b[35m\\u001b[1mLove OpenJSCAD? You can now donate to our open collective:\\u001b[22m\\u001b[39m\\n > \\u001b[34mhttps://opencollective.com/openjscad/donate\\u001b[0m')\""
   },
   "contributors": [
     {
@@ -35,10 +35,12 @@
   ],
   "license": "MIT",
   "dependencies": {
+    "@jscad/core": "1.6.1",
     "@jscad/csg": "0.3.6",
     "@jscad/io": "0.3.7",
     "@jscad/openscad-openjscad-translator": "0.0.10",
     "@jscad/scad-api": "^0.5.1",
+    "@jscad/examples": "1.6.1",
     "astring": "^1.0.2",
     "brace": "0.10.0",
     "esprima": "^3.1.3",
@@ -55,6 +57,7 @@
     "browserify": "14.3.0",
     "browserify-shim": "^3.8.12",
     "budo": "10.0.3",
+    "copy-dir": "^0.3.0",
     "minifyify": "^7.3.3"
   },
   "browserify": {
